Add liveness and readiness check endpoints, return 503 when not ready
