+++ src/edge_proxy/server.py
@@ -5,7 +5,7 @@ import structlog
 from fastapi import FastAPI, Header
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.middleware.gzip import GZipMiddleware
-from fastapi.responses import ORJSONResponse
+from fastapi.responses import ORJSONResponse, Response
 
 from edge_proxy.health_check.responses import HealthCheckResponse
 from fastapi_utils.tasks import repeat_every
@@ -40,11 +40,12 @@ async def unknown_key_error(request, exc):
 
 @app.get("/health", response_class=ORJSONResponse, deprecated=True)
 @app.get("/proxy/health", response_class=ORJSONResponse)
+@app.get("/proxy/health/readiness", response_class=ORJSONResponse)
 async def health_check():
     last_updated_at = environment_service.last_updated_at
     if not last_updated_at:
         return HealthCheckResponse(
-            status_code=500,
+            status_code=503,
             status="error",
             reason="environment document(s) not updated.",
             last_successful_update=None,
@@ -58,13 +59,18 @@ async def health_check():
         )
         if last_updated_at < threshold:
             return HealthCheckResponse(
-                status_code=500,
+                status_code=503,
                 status="error",
                 reason="environment document(s) stale.",
                 last_successful_update=last_updated_at,
             )
 
     return HealthCheckResponse(last_successful_update=last_updated_at)
+
+
+@app.get("/proxy/health/liveness")
+async def liveness_check():
+    return Response(status_code=200)
 
 
 @app.get("/api/v1/flags/", response_class=ORJSONResponse)
